% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_df.R
\name{read_df}
\alias{read_df}
\title{Function to read a data frame}
\usage{
read_df(fn, sep = "\\t", header = T, quote = "", row.names = T,
  stringsAsFactors = F, check.names = F, lessColInHeader = F,
  skip = 0)
}
\arguments{
\item{fn}{character. File name.}

\item{sep}{character. The filed seprator character. Default: a tab character.}

\item{quote}{character. The set of quote characters. Default: "".}

\item{row.names}{logical. Does each row has a name. Default: TRUE.}

\item{stringsAsFactors}{logical. Should characters should be converted to factors? Default: FALSE.}

\item{check.names}{logical. Should column names be checked to be valid names? Default: FALSE}

\item{lessColInHeader}{logial. Does the header line have one column less than the following lines? Default: FALSE.}

\item{skip}{integer. Number of lines to skip before beginning to read data. Default: 0.}
}
\value{
A data frame with data in the file.
}
\description{
Reads a file in table format and creates a data frame. It is very much like read.table() function, but it is faster than read.table().
}
\details{
The default values of this function are set to read a standard table with both row names and column names
Note: the function may not work properly if the parameters are not appropriate for the input file. So, please use the function cautiously.
}
\examples{
my_df = read_df('path/to/my/file')
}
