% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_feather.R
\name{read_feather_df}
\alias{read_feather_df}
\alias{write_feather_df}
\title{Functions to read or write a feather file.}
\usage{
read_feather_df(fn, columns = NULL, rownames.col = NULL)

write_feather_df(df, fn, rownames.title = NULL)
}
\arguments{
\item{fn}{character. File name.}

\item{columns}{integer or character vector. Columns (indexes or names) to read. Default: NULL to read all columns.}

\item{rownames.col}{integer or character. Column name or index containing row names. Default: NULL to avoid row names.}

\item{df}{data.frame to write.}

\item{rownames.title}{\code{NULL} or character. If \code{NULL}, row names are not written. 
Otherwise, row names are written as a column with title \code{rownames.title}.}
}
\value{
\code{read_feather_df} returns a data frame with data in the file. \code{write_feateher_df} returns nothing.
}
\description{
\code{read_feather_df} reads a feather file in a data frame. \code{write_feather_df} writes a data.frame to a feather file.
}
\examples{
my_df = read_feather_df(feather::feather_example("mtcars.feather"))
rownames(my_df) = paste0('car', 1:nrow(my_df))
write_feather_df(my_df, fn = "my_df.feather", rownames.title = "car_id")
my_df2 = read_feather_df("my_df.feather", rownames.col = "car_id")
}
