% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_print.R
\name{verbose_print}
\alias{verbose_print}
\title{Verbose print}
\usage{
verbose_print(msg, verbose = 1)
}
\arguments{
\item{msg}{character. message to print.}

\item{verbose}{numeric. print message if verbose > 0.}
}
\description{
This function prints message along with current time.
}
\details{
This function is useful for debugging. Whether the messages should be printed or not can be controlled by a single paramter verbose. If verbose = TRUE, the messages are printed, otherwise not. 
In addition, the function prints the current time which could be helpful to determine total runtime.
}
\examples{
verbose_print('this message will be printed with current time.', verbose = 1)
verbose_print('this message will NOT be printed.', verbose = 0)
}
