% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df.R
\name{write_df}
\alias{write_df}
\title{Function to write a data frame into a file}
\usage{
write_df(x, file, sep = "\\t", quote = F, row.names = T,
  col.names = NA)
}
\arguments{
\item{x}{data.frame}

\item{file}{either a character string or a connection open for writing.}

\item{sep}{character. The separator character.}

\item{quote}{logical. Should character or factor values be surrounded by double quotes?}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{col.names}{either a logical value indicating whether the column names of x are to be written along with x, or a character vector of column names to be written. If col.names = NA and row.names = TRUE a blank column name is added.}
}
\description{
This function is wrapper of the write.table() function with default values set to write a tab-delimited table with both row names and column names.
}
